django_template_proj/requirements.dev.txt
django
djangorestframework
celery[redis]
requests
google-auth-oauthlib
psycopg2